8337654: Relocate uncommon trap stub from SharedRuntime to OptoRuntime
