+++ src/hotspot/cpu/arm/runtime_arm.cpp
@@ -37,9 +37,145 @@
 #include "runtime/vframeArray.hpp"
 #include "utilities/globalDefinitions.hpp"
 #include "vmreg_arm.inline.hpp"
-#endif
 
 #define __ masm->
+
+//------------------------------generate_uncommon_trap_blob--------------------
+// Ought to generate an ideal graph & compile, but here's some ASM
+// instead.
+void OptoRuntime::generate_uncommon_trap_blob() {
+  // allocate space for the code
+  ResourceMark rm;
+
+  // setup code generation tools
+#ifdef _LP64
+  CodeBuffer buffer("uncommon_trap_blob", 2700, 512);
+#else
+  // Measured 8/7/03 at 660 in 32bit debug build
+  CodeBuffer buffer("uncommon_trap_blob", 2000, 512);
+#endif
+  // bypassed when code generation useless
+  MacroAssembler* masm               = new MacroAssembler(&buffer);
+  const Register Rublock = R6;
+  const Register Rsender = altFP_7_11;
+  assert_different_registers(Rublock, Rsender, Rexception_obj, R0, R1, R2, R3, R8, Rtemp);
+
+  //
+  // This is the entry point for all traps the compiler takes when it thinks
+  // it cannot handle further execution of compilation code. The frame is
+  // deoptimized in these cases and converted into interpreter frames for
+  // execution
+  // The steps taken by this frame are as follows:
+  //   - push a fake "unpack_frame"
+  //   - call the C routine Deoptimization::uncommon_trap (this function
+  //     packs the current compiled frame into vframe arrays and returns
+  //     information about the number and size of interpreter frames which
+  //     are equivalent to the frame which is being deoptimized)
+  //   - deallocate the "unpack_frame"
+  //   - deallocate the deoptimization frame
+  //   - in a loop using the information returned in the previous step
+  //     push interpreter frames;
+  //   - create a dummy "unpack_frame"
+  //   - call the C routine: Deoptimization::unpack_frames (this function
+  //     lays out values on the interpreter frame which was just created)
+  //   - deallocate the dummy unpack_frame
+  //   - return to the interpreter entry point
+  //
+  //  Refer to the following methods for more information:
+  //   - Deoptimization::uncommon_trap
+  //   - Deoptimization::unpack_frame
+
+  // the unloaded class index is in R0 (first parameter to this blob)
+
+  __ raw_push(FP, LR);
+  __ set_last_Java_frame(SP, FP, false, Rtemp);
+  __ mov(R2, Deoptimization::Unpack_uncommon_trap);
+  __ mov(R1, R0);
+  __ mov(R0, Rthread);
+  __ call(CAST_FROM_FN_PTR(address, Deoptimization::uncommon_trap));
+  __ mov(Rublock, R0);
+  __ reset_last_Java_frame(Rtemp);
+  __ raw_pop(FP, LR);
+
+#ifdef ASSERT
+  { Label L;
+    __ ldr_s32(Rtemp, Address(Rublock, Deoptimization::UnrollBlock::unpack_kind_offset()));
+    __ cmp_32(Rtemp, Deoptimization::Unpack_uncommon_trap);
+    __ b(L, eq);
+    __ stop("OptoRuntime::generate_uncommon_trap_blob: expected Unpack_uncommon_trap");
+    __ bind(L);
+  }
+#endif
+
+
+  // Set initial stack state before pushing interpreter frames
+  __ ldr_s32(Rtemp, Address(Rublock, Deoptimization::UnrollBlock::size_of_deoptimized_frame_offset()));
+  __ ldr(R2, Address(Rublock, Deoptimization::UnrollBlock::frame_pcs_offset()));
+  __ ldr(R3, Address(Rublock, Deoptimization::UnrollBlock::frame_sizes_offset()));
+
+  __ add(SP, SP, Rtemp);
+
+  // See if it is enough stack to push deoptimized frames.
+#ifdef ASSERT
+  // Compilers generate code that bang the stack by as much as the
+  // interpreter would need. So this stack banging should never
+  // trigger a fault. Verify that it does not on non product builds.
+  //
+  // The compiled method that we are deoptimizing was popped from the stack.
+  // If the stack bang results in a stack overflow, we don't return to the
+  // method that is being deoptimized. The stack overflow exception is
+  // propagated to the caller of the deoptimized method. Need to get the pc
+  // from the caller in LR and restore FP.
+  __ ldr(LR, Address(R2, 0));
+  __ ldr(FP, Address(Rublock, Deoptimization::UnrollBlock::initial_info_offset()));
+  __ ldr_s32(R8, Address(Rublock, Deoptimization::UnrollBlock::total_frame_sizes_offset()));
+  __ arm_stack_overflow_check(R8, Rtemp);
+#endif
+  __ ldr_s32(R8, Address(Rublock, Deoptimization::UnrollBlock::number_of_frames_offset()));
+  __ ldr_s32(Rtemp, Address(Rublock, Deoptimization::UnrollBlock::caller_adjustment_offset()));
+  __ mov(Rsender, SP);
+  __ sub(SP, SP, Rtemp);
+  //  __ ldr(FP, Address(FP));
+  __ ldr(FP, Address(Rublock, Deoptimization::UnrollBlock::initial_info_offset()));
+
+  // Push interpreter frames in a loop
+  Label loop;
+  __ bind(loop);
+  __ ldr(LR, Address(R2, wordSize, post_indexed));         // load frame pc
+  __ ldr(Rtemp, Address(R3, wordSize, post_indexed));      // load frame size
+
+  __ raw_push(FP, LR);                                     // create new frame
+  __ mov(FP, SP);
+  __ sub(Rtemp, Rtemp, 2*wordSize);
+
+  __ sub(SP, SP, Rtemp);
+
+  __ str(Rsender, Address(FP, frame::interpreter_frame_sender_sp_offset * wordSize));
+  __ mov(LR, 0);
+  __ str(LR, Address(FP, frame::interpreter_frame_last_sp_offset * wordSize));
+  __ subs(R8, R8, 1);                               // decrement counter
+  __ mov(Rsender, SP);
+  __ b(loop, ne);
+
+  // Re-push self-frame
+  __ ldr(LR, Address(R2));
+  __ raw_push(FP, LR);
+  __ mov(FP, SP);
+
+  // Call unpack_frames with proper arguments
+  __ mov(R0, Rthread);
+  __ mov(R1, Deoptimization::Unpack_uncommon_trap);
+  __ set_last_Java_frame(SP, FP, true, Rtemp);
+  __ call_VM_leaf(CAST_FROM_FN_PTR(address, Deoptimization::unpack_frames));
+  //  oop_maps->add_gc_map(__ pc() - start, new OopMap(frame_size_in_words, 0));
+  __ reset_last_Java_frame(Rtemp);
+
+  __ mov(SP, FP);
+  __ pop(RegisterSet(FP) | RegisterSet(PC));
+
+  masm->flush();
+  _uncommon_trap_blob = UncommonTrapBlob::create(&buffer, nullptr, 2 /* LR+FP */);
+}
 
 //------------------------------ generate_exception_blob ---------------------------
 // creates exception blob at the end
@@ -148,3 +284,6 @@ void OptoRuntime::generate_exception_blob() {
 
   _exception_blob = ExceptionBlob::create(&buffer, oop_maps, framesize_in_words);
 }
+
+#endif // COMPILER2
+
