+++ src/hotspot/share/opto/runtime.cpp
@@ -114,7 +114,8 @@ address OptoRuntime::_notify_jvmti_vthread_mount                  = nullptr;
 address OptoRuntime::_notify_jvmti_vthread_unmount                = nullptr;
 #endif
 
-ExceptionBlob* OptoRuntime::_exception_blob;
+UncommonTrapBlob*   OptoRuntime::_uncommon_trap_blob;
+ExceptionBlob*      OptoRuntime::_exception_blob;
 
 // This should be called in an assertion at the start of OptoRuntime routines
 // which are entered from compiled code (all of them)
@@ -138,6 +139,7 @@ static bool check_compiled_frame(JavaThread* thread) {
 
 bool OptoRuntime::generate(ciEnv* env) {
 
+  generate_uncommon_trap_blob();
   generate_exception_blob();
 
   // Note: tls: Means fetching the return oop out of the thread-local storage
